requirements.txt
requests==2.32.5
google-genai==1.45.0
python-dotenv==1.1.1